+++ Source/Examples/MonoMac/SimpleDemo/AppDelegate.designer.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleDemo
+{
+    [global::MonoMac.Foundation.Register ("AppDelegate")]
+    public partial class AppDelegate
+    {
+    }
+}