+++ Source/Examples/MonoMac/SimpleDemo/MainWIndow.Designer.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleDemo
+{
+    [global::MonoMac.Foundation.Register ("MainWindow")]
+    public partial class MainWindow
+    {
+    }
+}