+++ Source/Examples/WPF/SimpleDemo/Properties/AssemblyDescription.cs
@@ -1,0 +1,12 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="AssemblyDescription.cs" company="OxyPlot">
+//   Copyright (c) 2014 OxyPlot contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+using System.Reflection;
+using System.Windows;
+
+[assembly: AssemblyTitle("OxyPlot WPF demo")]
+[assembly: AssemblyDescription("")]
+[assembly: ThemeInfo(ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly)]