+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos.WinPhone/MainPage.xaml.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Navigation;
+using Microsoft.Phone.Controls;
+using Microsoft.Phone.Shell;
+
+namespace IssueDemos.WinPhone
+{
+    public partial class MainPage : global::Xamarin.Forms.Platform.WinPhone.FormsApplicationPage
+    {
+        public MainPage()
+        {
+            InitializeComponent();
+            SupportedOrientations = SupportedPageOrientation.PortraitOrLandscape;
+
+            global::Xamarin.Forms.Forms.Init();
+            OxyPlot.Xamarin.Forms.Platform.WP8.PlotViewRenderer.Init();
+            LoadApplication(new IssueDemos.App());
+        }
+    }
+}