+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos/DemoInfo.cs
@@ -1,0 +1,27 @@
+namespace IssueDemos
+{
+    using System;
+    using System.Reflection;
+    using Xamarin.Forms;
+
+    public class DemoInfo
+    {
+        private readonly DemoPageAttribute dpa;
+        private readonly Type type;
+
+        public DemoInfo(Type type)
+        {
+            this.type = type;
+            this.dpa = type.GetTypeInfo().GetCustomAttribute<DemoPageAttribute>();
+        }
+
+        public Page CreatePage()
+        {
+            return (Page)Activator.CreateInstance(this.type);
+        }
+
+        public string Title => dpa.Title;
+
+        public string Details => dpa.Details;
+    }
+}