+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos/DemoPageAttribute.cs
@@ -1,0 +1,15 @@
+namespace IssueDemos
+{
+    using System;
+
+    public sealed class DemoPageAttribute : Attribute
+    {
+        public DemoPageAttribute(string title, string details)
+        {
+            Title = title;
+            Details = details;
+        }
+        public string Title { get; private set; }
+        public string Details { get; private set; }
+    }
+}