+++ Source/OxyPlot.Wpf/Plot.Properties.cs
@@ -105,6 +105,13 @@ namespace OxyPlot.Wpf
                 "LegendItemSpacing", typeof(double), typeof(Plot), new PropertyMetadata(24.0, AppearanceChanged));
 
         /// <summary>
+        /// Identifies the <see cref="LegendLineSpacing"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty LegendLineSpacingProperty =
+            DependencyProperty.Register(
+                "LegendLineSpacing", typeof(double), typeof(Plot), new PropertyMetadata(0d, AppearanceChanged));
+
+        /// <summary>
         /// Identifies the <see cref="LegendMargin"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty LegendMarginProperty = DependencyProperty.Register(
@@ -231,8 +238,8 @@ namespace OxyPlot.Wpf
             typeof(IList<Color>),
             typeof(Plot),
             new PropertyMetadata(
-                new[] 
-            {  
+                new[]
+            {
                 Color.FromRgb(0x4E, 0x9A, 0x06),
                     Color.FromRgb(0xC8, 0x8D, 0x00),
                     Color.FromRgb(0xCC, 0x00, 0x00),
@@ -529,6 +536,23 @@ namespace OxyPlot.Wpf
         }
 
         /// <summary>
+        /// Gets or sets the spacing between columns of legend items (only for vertical orientation).
+        /// </summary>
+        /// <value>The spacing in device independent units.</value>
+        public double LegendColumnSpacing
+        {
+            get
+            {
+                return (double)this.GetValue(LegendColumnSpacingProperty);
+            }
+
+            set
+            {
+                this.SetValue(LegendColumnSpacingProperty, value);
+            }
+        }
+
+        /// <summary>
         /// Gets or sets LegendFont.
         /// </summary>
         public string LegendFont
@@ -609,8 +633,9 @@ namespace OxyPlot.Wpf
         }
 
         /// <summary>
-        /// Gets or sets LegendItemSpacing.
+        /// Gets or sets the horizontal spacing between legend items when the orientation is horizontal.
         /// </summary>
+        /// <value>The horizontal distance between items in device independent units.</value>
         public double LegendItemSpacing
         {
             get
@@ -625,6 +650,23 @@ namespace OxyPlot.Wpf
         }
 
         /// <summary>
+        /// Gets or sets the vertical spacing between legend items.
+        /// </summary>
+        /// <value>The spacing in device independent units.</value>
+        public double LegendLineSpacing
+        {
+            get
+            {
+                return (double)this.GetValue(LegendLineSpacingProperty);
+            }
+
+            set
+            {
+                this.SetValue(LegendLineSpacingProperty, value);
+            }
+        }
+
+        /// <summary>
         /// Gets or sets the max width of the legends.
         /// </summary>
         /// <value>The max width of the legends.</value>
@@ -632,12 +674,12 @@ namespace OxyPlot.Wpf
         {
             get
             {
-                return (double)GetValue(LegendMaxWidthProperty);
+                return (double)this.GetValue(LegendMaxWidthProperty);
             }
 
             set
             {
-                SetValue(LegendMaxWidthProperty, value);
+                this.SetValue(LegendMaxWidthProperty, value);
             }
         }
 
@@ -671,16 +713,6 @@ namespace OxyPlot.Wpf
             {
                 this.SetValue(LegendOrientationProperty, value);
             }
-        }
-
-        /// <summary>
-        /// Gets or sets the legend column spacing.
-        /// </summary>
-        /// <value>The legend column spacing.</value>
-        public double LegendColumnSpacing
-        {
-            get { return (double)GetValue(LegendColumnSpacingProperty); }
-            set { SetValue(LegendColumnSpacingProperty, value); }
         }
 
         /// <summary>