+++ Source/OxyPlot.Xamarin.Android/PlotView.cs
@@ -321,14 +321,13 @@ namespace OxyPlot.Xamarin.Android
                 return;
             }
 
-            if (!actualModel.Background.IsUndefined())
+            if (actualModel.Background.IsVisible())
             {
                 canvas.DrawColor(actualModel.Background.ToColor());
             }
             else
             {
-                // Use white as default background color
-                canvas.DrawColor(Color.White);
+                // do nothing
             }
 
             lock (this.invalidateLock)
@@ -349,13 +348,8 @@ namespace OxyPlot.Xamarin.Android
                 }
 
                 this.rc.SetTarget(canvas);
-                using (var bounds = new Rect())
-                {
-                    canvas.GetClipBounds(bounds);
-                    var width = bounds.Right - bounds.Left;
-                    var height = bounds.Bottom - bounds.Top;
-                    ((IPlotModel)actualModel).Render(this.rc, width / Scale, height / Scale);
-                }
+                
+                ((IPlotModel)actualModel).Render(this.rc, Width / Scale, Height / Scale);
             }
         }