+++ Source/OxyPlot.Xamarin.Forms.Platform.WP8/Forms.cs
@@ -9,6 +9,8 @@
 
 namespace OxyPlot.Xamarin.Forms.Platform.WP8
 {
+    using System;
+
     /// <summary>
     /// Initializes OxyPlot renderers for use with Xamarin.Forms.
     /// </summary>
@@ -17,11 +19,10 @@ namespace OxyPlot.Xamarin.Forms.Platform.WP8
         /// <summary>
         /// Initializes OxyPlot for Xamarin.Forms.
         /// </summary>
-        /// <remarks>This method must be called before Forms.Init().</remarks>
+        [Obsolete("Use PlotViewRenderer.Init() instead.")]
         public static void Init()
         {
-            // Just bring this assembly into the current appdomain.
-            // Forms.Init() should now find it!
+            PlotViewRenderer.Init();
         }
     }
 }