+++ Source/OxyPlot.Xamarin.Forms.Platform.iOS/PlotViewRenderer.cs
@@ -16,10 +16,28 @@ namespace OxyPlot.Xamarin.Forms.Platform.iOS
     public class PlotViewRenderer : ViewRenderer<Xamarin.Forms.PlotView, PlotView>
     {
         /// <summary>
-        /// Default constructor.
+        /// Initializes static members of the <see cref="PlotViewRenderer"/> class.
+        /// </summary>
+        static PlotViewRenderer()
+        {
+            Init();
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PlotViewRenderer"/> class.
         /// </summary>
         public PlotViewRenderer()
         {
+            // Do not delete
+        }
+
+        /// <summary>
+        /// Initializes the renderer.
+        /// </summary>
+        /// <remarks>This method must be called before a <see cref="T:PlotView" /> is used.</remarks>
+        public static new void Init()
+        {
+            OxyPlot.Xamarin.Forms.PlotView.IsRendererInitialized = true;
         }
 
         /// <summary>
@@ -37,13 +55,9 @@ namespace OxyPlot.Xamarin.Forms.Platform.iOS
             var plotView = new PlotView
             {
                 Model = this.Element.Model,
-                Controller = this.Element.Controller
+                Controller = this.Element.Controller,
+                BackgroundColor = this.Element.BackgroundColor.ToOxyColor().ToUIColor()
             };
-
-            if (this.Element.Model != null && this.Element.Model.Background.IsVisible())
-            {
-                plotView.BackgroundColor = this.Element.Model.Background.ToUIColor();
-            }
 
             this.SetNativeControl(plotView);
         }
@@ -69,6 +83,15 @@ namespace OxyPlot.Xamarin.Forms.Platform.iOS
             if (e.PropertyName == Xamarin.Forms.PlotView.ControllerProperty.PropertyName)
             {
                 this.Control.Controller = this.Element.Controller;
+            }
+
+            if (e.PropertyName == VisualElement.BackgroundColorProperty.PropertyName)
+            {
+                this.Control.BackgroundColor = this.Element.BackgroundColor.ToOxyColor().ToUIColor();
+            }
+
+            if (e.PropertyName == VisualElement.IsVisibleProperty.PropertyName) {
+                this.Control.InvalidatePlot (false);
             }
         }
     }